Modify the click listeners
Switch from a “click” listener to an on(“click”) listener. The click()
listener only adds listeners for existing elements. on() adds listeners
for all *potential future* elements

Change the listeners to an element that will not change after page
load, in this case that means the <ul>

Add in an argument that details which element within the parent will
trigger action. Either <span> for removing element, or <li> for cross
through styling.